% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FData.R
\docType{class}
\name{FData-class}
\alias{FData-class}
\alias{FData}
\title{A Reference Class which represents functional data.}
\description{
FData is a reference class which represents general independent and
identically distributed (i.i.d.) functional objects. The data can be ordered
by time (functional time series). In the last case, the field \code{X} represents
the time.
}
\section{Fields}{

\describe{
\item{\code{X}}{Numeric vector of length \emph{m} representing the
covariates/inputs.}

\item{\code{Y}}{Matrix of size \eqn{(n, m)} representing the observed
responses/outputs. \code{Y} consists of \emph{n} functions of \code{X} observed at
points \eqn{1,\dots,m}.}
}}


