% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMixHMM.R
\docType{class}
\name{ModelMixHMM-class}
\alias{ModelMixHMM-class}
\alias{ModelMixHMM}
\title{A Reference Class which represents a fitted MixHMM model.}
\description{
ModelMixHMM represents an estimated MixHMM model.
}
\section{Fields}{

\describe{
\item{\code{param}}{A \link{ParamMixHMM} object. It contains the
estimated values of the parameters.}

\item{\code{stat}}{A \link{StatMixHMM} object. It contains all the
statistics associated to the MixHMM model.}
}}

\section{Methods}{

\describe{
\item{\code{plot(what = c("clustered", "smoothed", "loglikelihood"), ...)}}{Plot method
\describe{
  \item{\code{what}}{The type of graph requested:
    \itemize{
      \item \code{"clustered" = } Clustered curves (field
        \code{klas} of class \link{StatMixHMM}).
      \item \code{"smoothed" = } Smoothed signal (field
        \code{smoothed} of class {StatMixHMM}).
      \item \code{"loglikelihood" = } Value of the log-likelihood for
        each iteration (field \code{stored_loglik} of class
        \link{StatMixHMM}).
    }
  }
  \item{\code{\dots}}{Other graphics parameters.}
}}

\item{\code{summary(digits = getOption("digits"))}}{Summary method.
\describe{
  \item{\code{digits}}{The number of significant digits to use when
    printing.}
}}
}}

\seealso{
\link{ParamMixHMM}, \link{StatMixHMM}
}
