% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMixRHLP.R
\docType{class}
\name{ModelMixRHLP-class}
\alias{ModelMixRHLP-class}
\alias{ModelMixRHLP}
\title{A Reference Class which represents a fitted MixRHLP model.}
\description{
ModelMixRHLP represents an estimated MixRHLP model.
}
\section{Fields}{

\describe{
\item{\code{param}}{A \link{ParamMixRHLP} object. It contains the
estimated values of the parameters.}

\item{\code{stat}}{A \link{StatMixRHLP} object. It contains all the
statistics associated to the MixRHLP model.}
}}

\section{Methods}{

\describe{
\item{\code{plot(what = c("estimatedsignal", "regressors", "loglikelihood"), ...)}}{Plot method.
\describe{
  \item{\code{what}}{The type of graph requested:
    \itemize{
      \item \code{"estimatedsignal" = } Estimated signal (field
        \code{Ex} of class \link{StatMixRHLP}).
      \item \code{"regressors" = } Polynomial regression components
        (fields \code{polynomials} and \code{pi_jgk} of class
        \link{StatMixRHLP}).
      \item \code{"loglikelihood" = } Value of the log-likelihood for
        each iteration (field \code{stored_loglik} of class
        \link{StatMixRHLP}).
    }
  }
  \item{\code{\dots}}{Other graphics parameters.}
}
By default, all the above graphs are produced.}

\item{\code{summary(digits = getOption("digits"))}}{Summary method.
\describe{
  \item{\code{digits}}{The number of significant digits to use when
    printing.}
}}
}}

\seealso{
\link{ParamMixRHLP}, \link{StatMixRHLP}
}
