% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatMixRHLP.R
\docType{class}
\name{StatMixRHLP-class}
\alias{StatMixRHLP-class}
\alias{StatMixRHLP}
\title{A Reference Class which contains statistics of a MixRHLP model.}
\description{
StatMixRHLP contains all the statistics associated to a
\link[=ParamMixRHLP]{MixRHLP} model.
}
\section{Fields}{

\describe{
\item{\code{pi_jgk}}{Array of size \eqn{(nm, R, K)} representing the logistic
proportion for cluster g.}

\item{\code{h_ig}}{Matrix of size \eqn{(n, K)} giving the posterior probabilities
(fuzzy segmentation matrix) that the curve \eqn{Y_{i}} originates from the
\eqn{g}-th RHLP model.}

\item{\code{c_ig}}{Hard segmentation logical matrix of dimension \eqn{(n, K)}
obtained by the Maximum a posteriori (MAP) rule: \eqn{c\_ig = 1 \
  \textrm{if} \ c\_ig = \textrm{arg} \ \textrm{max}_{s} \ h\_is;\ 0 \
  \textrm{otherwise}}{c_ig = 1 if c_ig = arg max_s h_is; 0 otherwise}, \eqn{g
  = 1,\dots,K}.}

\item{\code{klas}}{Column matrix of the labels issued from \code{c_ig}. Its elements are
\eqn{klas(i) = g}, \eqn{i = 1,\dots,n}.}

\item{\code{tau_ijgk}}{Array of size \eqn{(nm, R, K)} giving the posterior
probabilities that the observation \eqn{Y_{ij}} originates from the
\eqn{R}-th regime of the \eqn{g}-th RHLP model.}

\item{\code{polynomials}}{Array of size \eqn{(m, R, K)} giving the values of the
estimated polynomial regression components.}

\item{\code{weighted_polynomials}}{Array of size \eqn{(m, R, K)} giving the values
of the estimated polynomial regression components weighted by the prior
probabilities \code{pi_jgk}.}

\item{\code{Ex}}{Matrix of size \emph{(m, K)}. \code{Ex} is the curve expectation
(estimated signal): sum of the polynomial components weighted by the
logistic probabilities \code{pi_jgk}.}

\item{\code{loglik}}{Numeric. Observed-data log-likelihood of the MixRHLP model.}

\item{\code{com_loglik}}{Numeric. Complete-data log-likelihood of the MixRHLP model.}

\item{\code{stored_loglik}}{Numeric vector. Stored values of the log-likelihood at
each EM iteration.}

\item{\code{stored_com_loglik}}{Numeric vector. Stored values of the Complete
log-likelihood at each EM iteration.}

\item{\code{BIC}}{Numeric. Value of BIC (Bayesian Information Criterion).}

\item{\code{ICL}}{Numeric. Value of ICL (Integrated Completed Likelihood).}

\item{\code{AIC}}{Numeric. Value of AIC (Akaike Information Criterion).}

\item{\code{log_fg_xij}}{Matrix of size \eqn{(n, K)} giving the values of the
probability density function \eqn{f(y_{i} | h_ig = 1, \boldsymbol{x},
  \boldsymbol{\Psi})}{f(y_{i} | h_ig = 1, x, \Psi)}, \eqn{i = 1,\dots,n}.}

\item{\code{log_alphag_fg_xij}}{Matrix of size \eqn{(n, K)} giving the values of the
logarithm of the joint probability density function \eqn{f(y_{i}, \ h_{i} =
  g | \boldsymbol{x}, \boldsymbol{\Psi})}{f(y_{i}, h_{i} = g | x, \Psi)},
\eqn{i = 1,\dots,n}.}

\item{\code{log_tau_ijgk}}{Array of size \eqn{(nm, R, K)} giving the logarithm of
\code{tau_ijgk}.}
}}

\section{Methods}{

\describe{
\item{\code{computeStats(paramMixRHLP)}}{Method used in the EM algorithm to compute statistics based on
parameters provided by the object \code{paramMixRHLP} of class
\link{ParamMixRHLP}.}

\item{\code{CStep(reg_irls)}}{Method used in the CEM algorithm to update statistics.}

\item{\code{EStep(paramMixRHLP)}}{Method used in the EM algorithm to update statistics based on parameters
provided by the object \code{paramMixRHLP} of class \link{ParamMixRHLP}
(prior and posterior probabilities).}

\item{\code{MAP()}}{MAP calculates values of the fields \code{c_ig} and \code{klas}
      by applying the Maximum A Posteriori Bayes allocation rule.

      \eqn{c\_ig = 1 \ \textrm{if} \ c\_ig = \textrm{arg} \
      \textrm{max}_{s} \ h\_is;\ 0 \ \textrm{otherwise}
      }{c_ig = 1 if c_ig = arg max_s h_is; 0 otherwise}, \eqn{g = 1,\dots,K}.}
}}

\seealso{
\link{ParamMixRHLP}
}
