% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-toydataset.R
\docType{data}
\name{toydataset}
\alias{toydataset}
\title{A dataset composed of simulated time series with regime changes.}
\format{A data frame with 350 rows and 31 variables:
\describe{
\item{x}{The covariate variable which is the time in that case.}
\item{y1}{Times series with a wave form shape and for which a normally
distributed random noise has been added.}
\item{y2}{Same as \code{y1}.}
\item{y3}{Same as \code{y1}.}
\item{y4}{Same as \code{y1}.}
\item{y5}{Same as \code{y1}.}
\item{y6}{Same as \code{y1}.}
\item{y7}{Same as \code{y1}.}
\item{y8}{Same as \code{y1}.}
\item{y9}{Same as \code{y1}.}
\item{y10}{Same as \code{y1}.}
\item{y11}{Time series generated as follows:
\itemize{
\item  First regime: 120 values of Normally distributed random numbers
with mean 5 and variance 1.
\item Second regime: 70 values of Normally distributed random numbers
with mean 7 and variance 1.
\item Third regime: 160 values of Normally distributed random numbers
with mean 5 variance 1.
}
}
\item{y12}{Same as \code{y11}.}
\item{y13}{Same as \code{y11}.}
\item{y14}{Same as \code{y11}.}
\item{y15}{Same as \code{y11}.}
\item{y16}{Same as \code{y11}.}
\item{y17}{Same as \code{y11}.}
\item{y18}{Same as \code{y11}.}
\item{y19}{Same as \code{y11}.}
\item{y20}{Same as \code{y11}.}
\item{y21}{Time series generated as follows:
\itemize{
\item  First regime: 80 values of Normally distributed random numbers
with mean 7 variance 1.
\item Second regime: 130 values of Normally distributed random numbers
with mean 5 variance 1.
\item Third regime: 140 values of Normally distributed random numbers
with mean 4 variance 1.
}
}
\item{y22}{Same as \code{y21}.}
\item{y23}{Same as \code{y21}.}
\item{y24}{Same as \code{y21}.}
\item{y25}{Same as \code{y21}.}
\item{y26}{Same as \code{y21}.}
\item{y27}{Same as \code{y21}.}
\item{y28}{Same as \code{y21}.}
\item{y29}{Same as \code{y21}.}
\item{y30}{Same as \code{y21}.}
}}
\usage{
toydataset
}
\description{
A dataset composed of 30 simulated time series with regime changes.
}
\keyword{datasets}
